Fix: Comment 반영 및 컨트롤러, 서비스 기능 분리
