+++ src/main/java/com/nhnacademy/marketgg/auth/service/AuthInfoService.java
@@ -1,0 +1,8 @@
+package com.nhnacademy.marketgg.auth.service;
+
+import com.nhnacademy.marketgg.auth.dto.request.UpdateRequest;
+
+public interface AuthInfoService {
+    void update(UpdateRequest updateRequest);
+
+}
