+++ src/test/java/com/nhnacademy/marketgg/auth/controller/SignUpControllerTest.java
@@ -8,7 +8,7 @@ import com.nhnacademy.marketgg.auth.dto.response.ExistEmailResponse;
 import com.nhnacademy.marketgg.auth.dto.response.SignUpResponse;
 import com.nhnacademy.marketgg.auth.exception.EmailOverlapException;
 import com.nhnacademy.marketgg.auth.jwt.TokenUtils;
-import com.nhnacademy.marketgg.auth.service.AuthService;
+import com.nhnacademy.marketgg.auth.service.SignUpService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,22 +24,20 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(AuthController.class)
+@WebMvcTest(SignUpController.class)
 @Import(WebSecurityConfig.class)
 @MockBean({
         AuthenticationManager.class,
         TokenUtils.class,
         RedisTemplate.class
 })
-class AuthControllerTest {
+class SignUpControllerTest {
 
     @Autowired
     MockMvc mockMvc;
@@ -48,7 +46,7 @@ class AuthControllerTest {
     ObjectMapper mapper;
 
     @MockBean
-    AuthService authService;
+    SignUpService signUpService;
 
     @MockBean
     UserDetailsService userDetailsService;
@@ -67,7 +65,7 @@ class AuthControllerTest {
         ReflectionTestUtils.setField(testSignUpRequest, "name", "testName");
         ReflectionTestUtils.setField(testSignUpRequest, "phoneNumber", "010-1234-1234");
 
-        when(authService.signup(testSignUpRequest)).thenReturn(any(SignUpResponse.class));
+        when(signUpService.signup(testSignUpRequest)).thenReturn(any(SignUpResponse.class));
 
         mockMvc.perform(post("/auth/signup")
                                 .contentType(APPLICATION_JSON)
@@ -84,7 +82,7 @@ class AuthControllerTest {
         ReflectionTestUtils.setField(testEmailRequest, "email", "testEmail");
         ReflectionTestUtils.setField(testEmailRequest, "isReferrer", true);
 
-        when(authService.checkEmail(testEmailRequest))
+        when(signUpService.checkEmail(testEmailRequest))
                         .thenReturn(any(ExistEmailResponse.class));
 
         mockMvc.perform(post("/auth/check/email")
@@ -103,7 +101,7 @@ class AuthControllerTest {
         ReflectionTestUtils.setField(emailRequest, "email", "testEmail");
         ReflectionTestUtils.setField(emailRequest, "isReferrer", false);
 
-        when(authService.checkEmail(emailRequest))
+        when(signUpService.checkEmail(emailRequest))
                 .thenThrow(new EmailOverlapException(emailRequest.getEmail()));
 
         mockMvc.perform(post("/auth/check/email")
