+++ src/test/java/com/nhnacademy/marketgg/auth/service/impl/DefaultSignUpServiceTest.java
@@ -38,10 +38,10 @@ import static org.mockito.Mockito.*;
 @Import({
         WebSecurityConfig.class
 })
-class DefaultAuthServiceTest {
+class DefaultSignUpServiceTest {
 
     @InjectMocks
-    DefaultAuthService authService;
+    DefaultSignUpService defaultSignUpService;
 
     @Mock
     AuthRepository authRepository;
@@ -99,7 +99,7 @@ class DefaultAuthServiceTest {
 
         given(authRoleRepository.save(any(AuthRole.class))).willReturn(authRole);
 
-        authService.signup(testSignUpRequest);
+        defaultSignUpService.signup(testSignUpRequest);
 
         verify(authRepository, times(1)).save(any(auth.getClass()));
         // getDeclaringClass() 메서드는 이 클래스의 선언 클래스를 가져오는 데 사용됨.
@@ -120,7 +120,7 @@ class DefaultAuthServiceTest {
         ReflectionTestUtils.setField(testEmailRequest, "email", "test@test.com");
         ReflectionTestUtils.setField(testEmailRequest, "isReferrer", false);
 
-        authService.checkEmail(testEmailRequest);
+        defaultSignUpService.checkEmail(testEmailRequest);
 
         verify(authRepository, times(1)).existsByEmail(any());
     }
@@ -136,7 +136,7 @@ class DefaultAuthServiceTest {
         ReflectionTestUtils.setField(testEmailRequest, "isReferrer", false);
 
 
-        assertThatThrownBy(() -> authService.checkEmail(testEmailRequest))
+        assertThatThrownBy(() -> defaultSignUpService.checkEmail(testEmailRequest))
                 .isInstanceOf(EmailOverlapException.class);
 
         verify(authRepository, times(1)).existsByEmail(any());
